Rename directory to models so that sub-directories are in plural
