+++ lib/models/note_model.dart
